Switchover to MDI Interface Instead of Custom
Switched from a Custom Drawn Camera Interface over to a MDI Parent/Child
Interface. This will ease the custom drawing and allow a smoother and
more enjoyable experience overall for the end user as well as alleviate
development pains in the future as a MDI interface is easier to work
with in the end.